Move DeferQueue from util to vk
