+++ lib/vk/include/facade/vk/defer.hpp
@@ -1,6 +1,6 @@
 #pragma once
-#include <facade/util/defer_queue.hpp>
 #include <facade/util/unique.hpp>
+#include <facade/vk/defer_queue.hpp>
 
 namespace facade {
 template <typename T>
